+++ src/main/kotlin/com/dgu/prompt/blaybus_backend/SyncScheduler.kt
@@ -1,19 +1,18 @@
 package com.dgu.prompt.blaybus_backend
 
-import com.dgu.prompt.blaybus_backend.service.GoogleSheetsService
-import org.springframework.boot.CommandLineRunner
+import com.dgu.prompt.blaybus_backend.googleSheetsService.UsersSheetsService
 import org.springframework.scheduling.annotation.Scheduled
 import org.springframework.stereotype.Component
 
 @Component
 class SyncScheduler(
-    private val googleSheetsService: GoogleSheetsService
+    private val usersSheetsService: UsersSheetsService
 ) {
     @Scheduled(cron = "0 0 0 * * *") // 매 자정에 실행
 
 //    @Scheduled(cron = "0 37 20 * * *") // 오늘 오전 10시 31분에 실행
     fun syncGoogleSheetsData() {
-        googleSheetsService.syncData()
+        usersSheetsService.syncData()
     }
 }
 //구글 시트 바로 동기화
